rust: suricata-core crate
A zero-dependency crate, expecially for use by plugins.
