+++ .github/workflows/rust.yml
@@ -98,6 +98,8 @@ jobs:
         working-directory: rust/suricatasc
       - run: cargo fmt --check
         working-directory: rust/sys
+      - run: cargo fmt --check
+        working-directory: rust/core
       - name: Check if Cargo.lock.in is up to date
         run: |
           cp rust/Cargo.lock rust/Cargo.lock.in
