+++ rust/core/Makefile.am
@@ -1,0 +1,3 @@
+EXTRA_DIST =	Cargo.toml
+
+all-local: Cargo.toml
