+++ rust/src/bittorrent_dht/parser.rs
@@ -446,7 +446,7 @@ pub fn parse_bittorrent_dht_packet(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::direction::Direction;
+    use suricata_core::direction::Direction;
     use test_case::test_case;
 
     #[test_case(
