+++ rust/src/dcerpc/dcerpc_udp.rs
@@ -21,7 +21,7 @@ use crate::dcerpc::dcerpc::{
     DCERPCTransaction, DCERPC_MAX_TX, DCERPC_TYPE_REQUEST, DCERPC_TYPE_RESPONSE, PFCL1_FRAG, PFCL1_LASTFRAG,
     get_alstate_progress, ALPROTO_DCERPC, PARSER_NAME,
 };
-use crate::direction::{Direction, DIR_BOTH};
+use suricata_core::direction::{Direction, DIR_BOTH};
 use crate::flow::Flow;
 use nom7::Err;
 use suricata_sys::sys::AppProto;
