+++ rust/src/debug.rs
@@ -239,22 +239,7 @@ macro_rules! debug_validate_bug_on (
   };
 );
 
-#[cfg(not(feature = "debug-validate"))]
-#[macro_export]
-macro_rules! debug_validate_fail (
-  ($msg:expr) => {};
-);
-
-#[cfg(feature = "debug-validate")]
-#[macro_export]
-macro_rules! debug_validate_fail (
-  ($msg:expr) => {
-    // Wrap in a conditional to prevent unreachable code warning in caller.
-    if true {
-      panic!($msg);
-    }
-  };
-);
+pub use suricata_core::debug_validate_fail;
 
 #[macro_export]
 macro_rules! unwrap_or_return (
