+++ rust/src/ldap/ldap.rs
@@ -20,7 +20,7 @@
 use crate::applayer::{self, *};
 use crate::conf::conf_get;
 use crate::core::*;
-use crate::direction::Direction;
+use suricata_core::direction::Direction;
 use crate::flow::Flow;
 use crate::frames::*;
 use nom7 as nom;
