+++ rust/src/smb/smb1.rs
@@ -19,7 +19,7 @@
  * - check all parsers for calls on non-SUCCESS status
  */
 
-use crate::direction::Direction;
+use suricata_core::direction::Direction;
 use crate::smb::smb::*;
 use crate::smb::dcerpc::*;
 use crate::smb::events::*;
